# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CollectionRow:
    properties:
      canDelete:
        type: boolean
      canUpload:
        type: boolean
      color:
        type: string
      encryptedKey:
        type: string
      encryptedKeyNonce:
        type: string
      encryptedName:
        type: string
      id:
        type: string
      isShared:
        type: boolean
      nameNonce:
        type: string
      ownerUserId:
        type: string
      parentCollectionId:
        type: string
      uploadQuotaBytes:
        type: integer
      uploadUsedBytes:
        type: integer
    type: object
...
  handlers.CreateCollectionRequest:
    properties:
      encryptedKey:
        type: string
      encryptedKeyNonce:
        type: string
      encryptedName:
        type: string
      nameNonce:
        type: string
      parentCollectionId:
        type: string
    type: object
  handlers.CreateCollectionResult:
    properties:
      id:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.PubkeyResponse:
    properties:
      publicKey:
        type: string
    type: object
...
  handlers.ShareCollectionRequest:
    properties:
      canDelete:
        type: boolean
      canUpload:
        type: boolean
      encryptedCollectionKey:
        type: string
      recipientUserId:
        type: string
      uploadQuotaBytes:
        type: integer
    type: object
  handlers.ShareFederatedRequest:
    properties:
      canDelete:
        type: boolean
      canUpload:
        type: boolean
      encryptedCollectionKey:
        type: string
      recipientServer:
        type: string
      recipientUsername:
        type: string
      uploadQuotaBytes:
        type: integer
    type: object
  handlers.ShareFederatedResult:
    properties:
      inviteToken:
        type: string
      inviteUrl:
        type: string
    type: object
...
  handlers.UpdateCollectionRequest:
    properties:
      encryptedName:
        type: string
      nameNonce:
        type: string
    type: object
  handlers.UpdateColorRequest:
    properties:
      color:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: Self-hosted, end-to-end encrypted file storage with federation. All
    file content and metadata are encrypted client-side; the server stores only ciphertext.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kutup API
  version: 1.0.0
paths:
...
  /collections:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.CollectionRow'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List collections
      tags:
      - Collections
    post:
      consumes:
      - application/json
      parameters:
      - description: Encrypted collection name and key
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCollectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateCollectionResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a collection
      tags:
      - Collections
  /collections/{id}:
    delete:
      parameters:
      - description: Collection UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a collection and all its files
      tags:
      - Collections
    get:
      parameters:
      - description: Collection UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CollectionRow'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a collection
      tags:
      - Collections
    put:
      consumes:
      - application/json
      parameters:
      - description: Collection UUID
        in: path
        name: id
        required: true
        type: string
      - description: New encrypted name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Rename a collection
      tags:
      - Collections
  /collections/{id}/color:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Collection UUID
        in: path
        name: id
        required: true
        type: string
      - description: Color name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateColorRequest'
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set collection color
      tags:
      - Collections
...
  /collections/{id}/share:
    post:
      consumes:
      - application/json
      description: The client encrypts the collection key to the recipient's NaCl
        box public key before submitting.
      parameters:
      - description: Collection UUID
        in: path
        name: id
        required: true
        type: string
      - description: Share details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ShareCollectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Share a collection with a local user
      tags:
      - Collections
  /collections/{id}/share-federated:
    post:
      consumes:
      - application/json
      parameters:
      - description: Collection UUID
        in: path
        name: id
        required: true
        type: string
      - description: Federated share details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ShareFederatedRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ShareFederatedResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Share a collection with a user on a remote Kutup instance
      tags:
      - Collections
  /collections/fed-pubkey:
    get:
      parameters:
      - description: Remote username
        in: query
        name: username
        required: true
        type: string
      - description: Remote Kutup server base URL
        in: query
        name: server
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PubkeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fetch a remote user's public key (for federated sharing)
      tags:
      - Collections
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"